feat(api): create organization learner serializer
