+++ api/lib/infrastructure/serializers/jsonapi/organization-learner-follow-up/organization-learner-serializer.js
@@ -3,7 +3,7 @@ const { Serializer } = require('jsonapi-serializer');
 module.exports = {
   serialize(organizationLearner) {
     return new Serializer('organization-learner', {
-      attributes: ['lastName', 'firstName'],
+      attributes: ['lastName', 'firstName', 'email', 'username', 'authenticationMethods', 'division', 'group'],
     }).serialize(organizationLearner);
   },
 };
